refactor: move selenium prompt to separate file.
